remove deprecated example; fix access to unaligned data
