+++ examples/feinfo/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "feinfo"
+version = "0.0.0"
+description = "frontend information"
+authors = []
+license = ""
+repository = ""
+edition = "2021"
+
+[dependencies]
+libdvb = { path = ".." }
