+++ examples/femon/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "femon"
+version = "0.0.0"
+description = "frontend monitor"
+authors = []
+license = ""
+repository = ""
+edition = "2021"
+
+[dependencies]
+libdvb = { path = ".." }
