+++ examples/femon/src/main.rs
@@ -4,12 +4,6 @@ use {
         thread,
     },
 
-    anyhow::{
-        bail,
-        Context,
-        Result,
-    },
-
     libdvb::{
         FeDevice,
         FeStatus,
@@ -17,25 +11,25 @@ use {
 };
 
 
-fn main() -> Result<()> {
+fn main()  {
     let mut args = std::env::args().skip(1);
 
     let adapter = match args.next() {
-        Some(v) => v.parse::<u32>().context("adapter number")?,
-        None => bail!("adapter number not defined"),
-    };
-
-    let device = match args.next() {
-        Some(v) => v.parse::<u32>().context("device number")?,
+        Some(v) => v.parse().unwrap(),
         None => 0,
     };
 
-    let fe = FeDevice::open_ro(adapter, device)?;
+    let device = match args.next() {
+        Some(v) => v.parse().unwrap(),
+        None => 0,
+    };
+
+    let fe = FeDevice::open_ro(adapter, device).unwrap();
     let mut status = FeStatus::default();
 
     let delay = Duration::from_secs(1);
     loop {
-        status.read(&fe)?;
+        status.read(&fe).unwrap();
         println!("{}", &status);
         thread::sleep(delay);
     }
