+++ src/fe/sys.rs
@@ -633,44 +633,54 @@ impl fmt::Debug for DtvProperty {
 
         match self.cmd {
             DTV_FREQUENCY => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_FREQUENCY");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_MODULATION => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_MODULATION");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_BANDWIDTH_HZ => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_BANDWIDTH_HZ");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_INVERSION => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_INVERSION");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_SYMBOL_RATE  => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_SYMBOL_RATE");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_INNER_FEC => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_INNER_FEC");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_PILOT => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_PILOT");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_ROLLOFF => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_ROLLOFF");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_DELIVERY_SYSTEM => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_DELIVERY_SYSTEM");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
             DTV_API_VERSION => {
+                let data = self.get_data();
                 s.field(FIELD_CMD, &"DTV_API_VERSION");
-                s.field(FIELD_DATA, unsafe { &self.u.data });
+                s.field(FIELD_DATA, &data);
             }
 
             /* Quality parameters */
@@ -728,6 +738,13 @@ impl DtvProperty {
             u: DtvPropertyData { data },
             result: 0,
         }
+    }
+
+    #[inline]
+    pub(crate) fn get_data(&self) -> u32 {
+        let u_ptr = std::ptr::addr_of!(self.u);
+        let u = unsafe { u_ptr.read_unaligned() };
+        unsafe { u.data }
     }
 }
