+++ src/lib.rs
@@ -2,16 +2,11 @@
 extern crate anyhow;
 
 
-pub mod ca;
 pub mod fe;
 pub mod net;
 
 
 pub use {
-    ca::{
-        CaDevice,
-    },
-
     fe::{
         FeDevice,
         FeStatus,
